# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API for managing contacts information",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Endpoints for managing contacts"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/debug/data": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Racheal"
        },
        "lastName": {
          "type": "string",
          "example": "Katono"
        },
        "email": {
          "type": "string",
          "example": "rafaelktn29@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Black"
        },
        "birthday": {
          "type": "string",
          "example": "1997-09-02"
        }
      }
    },
    "ContactInput": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Tom"
        },
        "lastName": {
          "type": "string",
          "example": "Seguja"
        },
        "email": {
          "type": "string",
          "example": "tomseguja@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Green"
        },
        "birthday": {
          "type": "string",
          "example": "1965-06-15"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ]
    },
    "ContactResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "682e23e8cc8f14903304a546"
        },
        "firstName": {
          "type": "string",
          "example": "Joan"
        },
        "lastName": {
          "type": "string",
          "example": "Nakibuule"
        },
        "email": {
          "type": "string",
          "example": "joanruth@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "White"
        },
        "birthday": {
          "type": "string",
          "example": "1970-03-22"
        }
      }
    },
    "NewContactResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "682e23e8cc8f14903304a544"
        },
        "message": {
          "type": "string",
          "example": "Contact created successfully"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Error message"
        },
        "message": {
          "type": "string",
          "example": "Detailed error information"
        }
      }
    }
  }
}